requirements.txt
filelock==3.19.1
fsspec==2025.10.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.90
packaging==26.0
pillow==11.3.0
protobuf==6.33.5
sympy==1.14.0
tensorboardX==2.6.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
